# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Linguo API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "description": "Authenticate a user with email and password. Returns an access token in the response body and sets a httpOnly refresh token cookie.",
                "operationId": "login_api_v1_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid email or password"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "User logout",
                "description": "Invalidates the refresh token stored in the httpOnly cookie and clears it. Returns 204 No Content regardless of whether the cookie was present.",
                "operationId": "logout_api_v1_auth_logout_post",
                "responses": {
                    "204": {
                        "description": "Successfully logged out"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token rotation",
                "description": "Rotates the refresh token: invalidates the current one, issues a new refresh token (set as httpOnly cookie) and returns a new access token in the response body. Requires a valid refresh token cookie.",
                "operationId": "refresh_api_v1_auth_refresh_post",
                "responses": {
                    "200": {
                        "description": "Tokens successfully refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Refresh token is missing, invalid or expired"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "User registration",
                "description": "Register a new user",
                "operationId": "register_api_v1_auth_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegistrationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Reset password",
                "description": "Change password for the authenticated user after old password verification.",
                "operationId": "reset_password_api_v1_auth_reset_password_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Password successfully changed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResetPasswordResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Password validation failed"
                    },
                    "401": {
                        "description": "Authentication required"
                    },
                    "403": {
                        "description": "Old password is incorrect"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccessTokenResponse": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "title": "Access Token"
                    },
                    "token_type": {
                        "type": "string",
                        "title": "Token Type",
                        "default": "bearer"
                    }
                },
                "type": "object",
                "required": [
                    "access_token"
                ],
                "title": "AccessTokenResponse"
            },
...
            "AuthenticationRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "maxLength": 255,
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password"
                ],
                "title": "AuthenticationRequest"
            },
            "AuthenticationResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success",
                        "default": true
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Authentication successful"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AccessTokenResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp"
                    }
                },
                "type": "object",
                "title": "AuthenticationResponse",
                "description": "Response schema for successful login."
            },
...
            "ErrorDetail": {
                "properties": {
                    "code": {
                        "type": "string",
                        "title": "Code",
                        "description": "Machine-readable error code (e.g. USER_NOT_FOUND).",
                        "examples": [
                            {
                                "AuthErrorCodes": {
                                    "EMAIL_OR_PASSWORD_INCORRECT": "EMAIL_OR_PASSWORD_INCORRECT",
                                    "FORBIDDEN": "FORBIDDEN",
                                    "INVALID_OLD_PASSWORD": "INVALID_OLD_PASSWORD",
                                    "INVALID_OR_EXPIRED_REFRESH_TOKEN": "INVALID_OR_EXPIRED_REFRESH_TOKEN",
                                    "NEW_PASSWORD_MUST_DIFFER": "NEW_PASSWORD_MUST_DIFFER"
                                },
                                "BaseErrorCodes": {
                                    "INVALID_FILTER": "INVALID_FILTER",
                                    "INVALID_PAGINATION": "INVALID_PAGINATION",
                                    "PERMISSION_DENIED": "PERMISSION_DENIED",
                                    "RESOURCE_NOT_FOUND": "RESOURCE_NOT_FOUND"
                                },
                                "CardErrorCodes": {
                                    "CARD_NOT_FOUND": "CARD_NOT_FOUND",
                                    "INVALID_FIELD": "INVALID_FIELD",
                                    "INVALID_FILE_TYPE": "INVALID_FILE_TYPE"
                                },
                                "DeckErrorCodes": {
                                    "DECK_ALREADY_EXISTS": "DECK_ALREADY_EXISTS",
                                    "DECK_BELONGS_TO_ANOTHER_USER": "DECK_BELONGS_TO_ANOTHER_USER",
                                    "DECK_NOT_FOUND": "DECK_NOT_FOUND",
                                    "INVALID_DECK_NAME": "INVALID_DECK_NAME"
                                },
                                "FilesErrorCodes": {
                                    "FILE_ACCESS_DENIED": "FILE_ACCESS_DENIED",
                                    "FILE_ALREADY_EXISTS": "FILE_ALREADY_EXISTS",
                                    "FILE_BELONGS_TO_ANOTHER_USER": "FILE_BELONGS_TO_ANOTHER_USER",
                                    "FILE_DELETE_FAILED": "FILE_DELETE_FAILED",
                                    "FILE_NOT_FOUND": "FILE_NOT_FOUND",
                                    "FILE_TOO_LARGE": "FILE_TOO_LARGE",
                                    "FILE_UPLOAD_FAILED": "FILE_UPLOAD_FAILED",
                                    "INVALID_FILE_TYPE": "INVALID_FILE_TYPE"
                                },
                                "ProfilesErrorCodes": {
                                    "BIO_MUST_BE_AT_MOST_300_CHARACTERS": "BIO_MUST_BE_AT_MOST_300_CHARACTERS",
                                    "PROFILE_NOT_FOUND": "PROFILE_NOT_FOUND",
                                    "USERNAME_ALREADY_IN_USE": "USERNAME_ALREADY_IN_USE",
                                    "USERNAME_MUST_BE_BETWEEN_3_AND_50_CHARACTERS": "USERNAME_MUST_BE_BETWEEN_3_AND_50_CHARACTERS"
                                },
                                "QuizErrorCodes": {
                                    "ALL_FIELDS_ARE_EXCLUDED": "ALL_FIELDS_ARE_EXCLUDED",
                                    "AT_LEAST_ONE_INPUT_TYPE_MUST_BE_AVAILABLE": "AT_LEAST_ONE_INPUT_TYPE_MUST_BE_AVAILABLE",
                                    "AT_LEAST_ONE_PRESENTATION_TYPE_MUST_BE_AVAILABLE": "AT_LEAST_ONE_INPUT_TYPE_MUST_BE_AVAILABLE",
                                    "IMAGE_DOES_NOT_EXIST_FOR_CARD": "IMAGE_DOES_NOT_EXIST_FOR_CARD",
                                    "INVALID_FIELD_TYPE": "INVALID_FIELD_TYPE",
                                    "INVALID_RATING": "INVALID_RATING",
                                    "NO_COMPATIBLE_COMBINATIONS_FOUND": "NO_COMPATIBLE_COMBINATIONS_FOUND",
                                    "SOUND_DOES_NOT_EXIST_FOR_CARD": "SOUND_DOES_NOT_EXIST_FOR_CARD",
                                    "THERE_IS_NO_CARD_TO_REVIEW": "THERE_IS_NO_CARD_TO_REVIEW"
                                },
                                "UserErrorCodes": {
                                    "EMAIL_ALREADY_IN_USE": "EMAIL_ALREADY_IN_USE",
                                    "USER_NOT_FOUND": "USER_NOT_FOUND"
                                }
                            }
                        ]
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Human-readable description of the error.",
                        "examples": [
                            "User with this email not found"
                        ]
                    },
                    "details": {
                        "title": "Details",
                        "description": "Optional structured payload with additional context (e.g. validation field errors)."
                    }
                },
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "title": "ErrorDetail",
                "description": "Inner object that describes what went wrong."
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success",
                        "default": false
                    },
                    "error": {
                        "$ref": "#/components/schemas/ErrorDetail"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Time",
                        "description": "Timestamp of when the error occurred (ISO 8601 format)."
                    }
                },
                "type": "object",
                "required": [
                    "error"
                ],
                "title": "ErrorResponse"
            },
...
            "RegistrationRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "maxLength": 255,
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 55,
                        "minLength": 3,
                        "title": "Username"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "password",
                    "username"
                ],
                "title": "RegistrationRequest"
            },
            "ResetPasswordRequest": {
                "properties": {
                    "old_password": {
                        "type": "string",
                        "title": "Old Password"
                    },
                    "new_password": {
                        "type": "string",
                        "title": "New Password"
                    },
                    "confirmed_password": {
                        "type": "string",
                        "title": "Confirmed Password"
                    }
                },
                "type": "object",
                "required": [
                    "old_password",
                    "new_password",
                    "confirmed_password"
                ],
                "title": "ResetPasswordRequest"
            },
            "ResetPasswordResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success",
                        "default": true
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "Password has been reset successfully"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data"
                    },
                    "timestamp": {
                        "type": "string",
                        "title": "Timestamp"
                    }
                },
                "type": "object",
                "title": "ResetPasswordResponse",
                "description": "Response schema for successful password reset."
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}